ENHANCED: arc: carry sub-pixel precision through geometry and rendering
points_arc, RedrawAreaArc, includeArrowsInAreaArc, computeArc and
the <-start/<-end getters now use double for the computed sx/sy/ex/ey
coordinates and pass them to attached arrows via toNum() without an
integer round-trip.  r_arc takes its bounding rectangle and angles
as double, so drawArcGraphical (->draw_arc) and the arc class no
longer truncate angles to int before handing them to cairo.  The
bounding area set on the arc is still rounded — out to floor/ceil
with the existing 1-pixel margin — because Area itself is integer.

+++ src/gra/proto.h
@@ -1,7 +1,8 @@
 #define COMMON(type) SO_LOCAL type
 
 /* /swi40/jan/src/pl/packages/xpce/src/gra/arc.c */
-COMMON(void)	points_arc(ArcObj a, int *sx, int *sy, int *ex, int *ey);
+COMMON(void)	points_arc(ArcObj a, double *sx, double *sy,
+			   double *ex, double *ey);
 COMMON(status)	makeClassArc(Class class);
 
 /* /swi40/jan/src/pl/packages/xpce/src/gra/arrow.c */
